Remove panel_job_instance rule
